+++ src/org/mozilla/mozstumbler/client/cellscanner/DefaultCellScanner.java
@@ -9,8 +9,8 @@ import android.telephony.CellSignalStrengthWcdma;
 import android.telephony.TelephonyManager;
 
 import org.mozilla.mozstumbler.service.AppGlobals;
-import org.mozilla.mozstumbler.service.scanners.cellscanner.CellInfo;
-import org.mozilla.mozstumbler.service.scanners.cellscanner.CellScannerNoWCDMA;
+import org.mozilla.mozstumbler.service.stumblerthread.scanners.cellscanner.CellInfo;
+import org.mozilla.mozstumbler.service.stumblerthread.scanners.cellscanner.CellScannerNoWCDMA;
 import java.util.List;
 
 public class DefaultCellScanner extends CellScannerNoWCDMA {
