+++ src/org/mozilla/mozstumbler/client/mapview/Searcher.java
@@ -6,10 +6,10 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
+
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.mozilla.mozstumbler.service.AbstractCommunicator;
+import org.mozilla.mozstumbler.service.utils.AbstractCommunicator;
 import org.mozilla.mozstumbler.service.AppGlobals;
 
 public class Searcher extends AbstractCommunicator {
