+++ src/org/mozilla/mozstumbler/service/utils/DateTimeUtils.java
@@ -16,7 +16,7 @@ public final class DateTimeUtils {
     private static final DateFormat sLocaleFormat = DateFormat.getDateTimeInstance(DateFormat.SHORT, DateFormat.SHORT);
     private static final DateFormat sISO8601Format = new SimpleDateFormat("yyyy-MM-dd", Locale.US);
 
-    public static final long MILLISECONDS_PER_DAY = 86400000; // milliseconds/day
+    public static final long MILLISECONDS_PER_DAY = 86400000;
 
     static {
         sISO8601Format.setTimeZone(TimeZone.getTimeZone("UTC"));
