+++ src/org/mozilla/mozstumbler/service/utils/Zipper.java
@@ -11,43 +11,40 @@ import java.util.zip.GZIPOutputStream;
 public class Zipper {
     public static byte[] zipData(byte[] data) throws IOException {
         final ByteArrayOutputStream os = new ByteArrayOutputStream();
-        GZIPOutputStream gstream = null;
+        GZIPOutputStream gstream = new GZIPOutputStream(os);
+        if (gstream == null) {
+            return null;
+        }
         byte[] output;
         try {
-            gstream = new GZIPOutputStream(os);
             gstream.write(data);
             gstream.finish();
             output = os.toByteArray();
         } finally {
+            gstream.close();
             os.close();
-            if (gstream != null) {
-                gstream.close();
-            }
         }
         return output;
     }
 
     public static String unzipData(byte[] data) throws IOException {
-        final ByteArrayInputStream bs = new ByteArrayInputStream(data);
-        GZIPInputStream gstream = null;
         String result = "";
+        final ByteArrayInputStream bs = new ByteArrayInputStream(data);
+        GZIPInputStream gstream = new GZIPInputStream(bs);
+        if (gstream == null) {
+            return result;
+        }
 
         try {
-            gstream = new GZIPInputStream(bs);
-
             InputStreamReader reader = new InputStreamReader(gstream);
             BufferedReader in = new BufferedReader(reader);
-
             String read;
             while ((read = in.readLine()) != null) {
                 result += read;
             }
-
         } finally {
+            gstream.close();
             bs.close();
-            if (gstream != null) {
-                gstream.close();
-            }
         }
         return result;
     }
